WIP: quic-be: Temporary fix to avoid triggering a BUG_ON() from dgram parser
At this time, the ->cids tree attached to the quic_conn object is empty during the
first QUIC server packets parsing. This leads haproxy to crash when it dereferences
the node returned by eb64_first().
Run the BUG_ON() statement only when ->cids is not empty (or when the node returned
by eb64_first() is not NULL).